Rename : Folder Structure Modify View와 Component의 구분이 모호하던 폴더 구조를 수정하였습니다.
