--- healog_front/src/components/account/trainer/AccountViewTrainer.vue
@@ -1,13 +1,0 @@
-<template>
-    <div>
-        <h3>Trainer</h3>
-    </div>
-</template>
-
-<script setup>
-
-</script>
-
-<style scoped>
-
-</style>
