+++ healog_front/src/components/diary/DiaryView.vue
@@ -1,0 +1,14 @@
+<template>
+    <div>
+        <h2>Diary Vue</h2>
+        <NavigationView />
+    </div>
+</template>
+
+<script setup>
+import NavigationView from '@/components/common/NavigationView.vue'
+</script>
+
+<style scoped>
+
+</style>
