+++ healog_front/src/router/index.js
@@ -1,19 +1,23 @@
-import { createRouter, createWebHistory } from 'vue-router'
-import AccountViewUser from '@/components/account/user/AccountViewUser.vue'
-import AccountViewTrainer from '@/components/account/trainer/AccountViewTrainer.vue'
-import AccountViewChoose from '@/components/account/AccountViewChoose.vue'
-import AccountViewUserLogin from '@/components/account/user/AccountViewUserLogin.vue'
-import AccountViewUserSignup from '@/components/account/user/AccountViewUserSignup.vue'
-import App from '@/App.vue'
-import { useUserStore } from '@/stores/user'
-
+import { createRouter, createWebHistory } from 'vue-router';
+import AccountViewUser from '@/views/account/user/AccountViewUser.vue';
+import AccountViewTrainer from '@/views/account/trainer/AccountViewTrainer.vue';
+import AccountViewChoose from '@/components/account/AccountViewChoose.vue';
+import AccountViewUserLogin from '@/views/account/user/AccountViewUserLogin.vue';
+import AccountViewUserSignup from '@/views/account/user/AccountViewUserSignup.vue';
+import App from '@/App.vue';
+import { useUserStore } from '@/stores/user';
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     {
       path: '/',
-      name: 'home',
+      name: 'home'
+    },
+    {
+      path: '/account',
+      name: 'account',
+      component: AccountViewChoose,
     },
     {
       path: '/user',
@@ -26,27 +30,17 @@ const router = createRouter({
       component: AccountViewTrainer,
     },
     {
-      path: '/account',
-      name: 'account',
-      component: AccountViewChoose
-    },
-    {
       path: '/user/login',
       name: 'userLogin',
-      component: AccountViewUserLogin
+      component: AccountViewUserLogin,
     },
     {
       path: '/user/signup',
       name: 'userSignup',
-      component: AccountViewUserSignup
-    }
+      component: AccountViewUserSignup,
+    },
   ],
-})
+});
 
 
-router.beforeEach((to, from)=>{
-  const store = useUserStore()
-  store.checkTokenValid()
-})
-
-export default router
+export default router;
