+++ healog_front/src/views/account/trainer/AccountViewTrainer.vue
@@ -1,0 +1,13 @@
+<template>
+    <div>
+        <h3>Trainer</h3>
+    </div>
+</template>
+
+<script setup>
+
+</script>
+
+<style scoped>
+
+</style>
